// src/config/response/response-code.enum.ts
export enum ResponseCode {
  // Success
  SUCCESS = 1000,

  // Common
  TOKEN_EMPTY = 2000,
  TOKEN_VERIFICATION_FAILURE = 2001,
  TOKEN_VERIFICATION_SUCCESS = 2002,

  // Validation
  VALIDATION_ERROR = 3000,
  MISSING_REQUIRED_FIELD = 3001,
  INVALID_INPUT = 3002,

  // Auth
  AUTH_UNAUTHORIZED = 4000,
  AUTH_FORBIDDEN = 4001,
  AUTH_INVALID_CREDENTIALS = 4002,
  AUTH_ACCOUNT_INACTIVE = 4003,
  AUTH_ACCOUNT_WITHDRAWN = 4004,

  // Server
  DB_ERROR = 5000,
  SERVER_ERROR = 5001,
  EXTERNAL_SERVICE_ERROR = 5002,
}